Rearranging the files to maintain consistency and fix minor issues in verification
